Revert "float64 -> float32"
This reverts commit da0f6a8f8de4a6900e13543d89f1b8ed3b6329c1.

+++ data.go
@@ -37,10 +37,10 @@ type WifiAccessPoint struct {
 
 type Response struct {
 	Location Point   `json:"location"` // The user’s estimated latitude and longitude, in degrees.
-	Accuracy float32 `json:"accuracy"` // The accuracy of the estimated location, in meters.
+	Accuracy float64 `json:"accuracy"` // The accuracy of the estimated location, in meters.
 }
 
 type Point struct {
-	Lat float32 `json:"lat"`
-	Lon float32 `json:"lng"`
+	Lat float64 `json:"lat"`
+	Lon float64 `json:"lng"`
 }
